Fixed and tested pet-interface: Pedro Italo

+++ codigo/src/domain/adoption/application/use-cases/delete-pet.js
@@ -1,6 +1,7 @@
-import { left } from '../../../../../core/Either.js';
+import { left, right } from '../../../../../core/Either.js';
 import { RequestMissingDataError } from '../errors/request-missing-data-error.js';
 import { ResourceNotFoundError } from '../../../../../core/errors/resource-not-found-error.js';
+import { NotAllowedError } from '../../../../../core/errors/not-allowed-error.js';
 
 
 /*
@@ -12,25 +13,36 @@ import { ResourceNotFoundError } from '../../../../../core/errors/resource-not-f
 
 export class DeletePetUseCase {
     petRepository;
-    constructor(petRepository) {
+    donorRepository;
+    constructor(petRepository, donorRepository) {
         this.petRepository = petRepository;
+        this.donorRepository = donorRepository;
     }
 
     async execute({
-       id
+       petId,
+       donorId
     }) {
 
-        if(!id) {
+        if(!petId || !donorId) {
             return left(new RequestMissingDataError())
         }
 
-        const {pet} = await this.petRepository.findById(id);
+        const {pet} = await this.petRepository.findById(petId);
         console.log(pet)
 
         if(!pet) {
             return left(new ResourceNotFoundError());
         }
 
+        const {donor} = await this.donorRepository.findById(donorId)
+
+        if(!donor || donor.id !== pet.donorId) {
+            return left(new NotAllowedError())
+        }
+
         await this.petRepository.delete(pet);
+
+        return right("Pet Deleted");
     }
 }
